featuresにquiz作成し、ロジックを映した
