+++ src/features/Quiz/Quiz.module.css
@@ -8,7 +8,6 @@
   padding: 20px;
   box-sizing: border-box;
 }
-
 .mainPanel {
   width: 100%;
   max-width: 600px;
@@ -147,10 +146,6 @@
   transition: background-color 0.2s;
 }
 
-.completeButton:hover {
-  background-color: #218838;
-}
-
 .resetButtonContainer {
   margin-top: 20px;
   width: 100%;
@@ -158,7 +153,7 @@
 }
 
 .resetButton {
-  background-color: #e91e63; /* ★ボタンの色を修正 */
+  background-color: #e91e63;
   color: white;
   border: none;
   padding: 12px 15px;
@@ -170,5 +165,5 @@
 }
 
 .resetButton:hover {
-  background-color: #c2185b; /* ★ホバー時の色も修正 */
+  background-color: #c2185b;
 }
